requirements.txt
pillow==9.0.0
reportlab==3.6.12
PyJWT==2.8.0
Flask==3.0.0
bcrypt==4.1.3
SQLAlchemy==2.0.30